GHI-316 remove pendingsafevc
