+++ SafeUIKit/SafeAppUI/en.lproj/Localizable.strings
@@ -220,15 +220,6 @@
 /* Title of abort safe creation alert */
 "pending_safe.abort_alert.title" = "Aborting Safe Creation";
 
-/* Address label */
-"pending_safe.address" = "Safe address";
-
-/* Balance label */
-"pending_safe.balanceLabel" = "Balance";
-
-/* Cancel safe creation button */
-"pending_safe.cancel" = "Cancel";
-
 /* Pending safe failed alert's message */
 "pending_safe.failed_alert.message" = "Error: %@ Please try again later.";
 
@@ -237,33 +228,6 @@
 
 /* Pending safe failed alert's title */
 "pending_safe.failed_alert.title" = "Error";
-
-/* Info label about safe creation */
-"pending_safe.info" = "Your safe is being created. Please follow status updates below.";
-
-/* Retry button title */
-"pending_safe.retry" = "Retry";
-
-/* Account received enough funds */
-"pending_safe.status.account_funded" = "Account has enough funds. Proceeding.";
-
-/* Deployment accepted by blockchain */
-"pending_safe.status.deployment_accepted" = "Creating Safe in the blockchain...";
-
-/* Deployment started status */
-"pending_safe.status.deployment_started" = "Safe creation started.";
-
-/* Deployment succeeded */
-"pending_safe.status.deployment_success" = "Successfully created new Safe in the blockchain";
-
-/* Error during safe creation. Retry later. */
-"pending_safe.status.error" = "Please retry later.";
-
-/* Not enough funds in account */
-"pending_safe.status.not_enough_funds" = "Not enough funds for safe creation: %@. Please transfer at least %@.";
-
-/* Title of pending safe screen */
-"pending_safe.title" = "Creating new safe";
 
 /* Safe Address label. */
 "safe_address.address" = "Address";
